# contracts/extensions/v1alpha1/provider-contracts.openapi.yaml
openapi: 3.1.0
info:
  title: Epydios AgentOps Control Plane Extension Provider Contracts
  version: v1alpha1
  summary: Versioned HTTP/JSON contracts for PolicyProvider, EvidenceProvider, and ProfileResolver.
  description: |
    This contract is the public plug-in surface used by OSS baseline providers and private
    providers (including AIMXS integrations). Breaking changes require a new version.
servers:
  - url: https://provider.internal
...
paths:
...
  /v1alpha1/policy-provider/evaluate:
    post:
      tags: [PolicyProvider]
      operationId: policyEvaluate
      summary: Evaluate a policy decision request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyEvaluateRequest"
      responses:
        "200":
          description: Policy evaluation decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyEvaluateResponse"
        "400":
          $ref: "#/components/responses/ProviderError"
        "401":
          $ref: "#/components/responses/ProviderError"
        "403":
          $ref: "#/components/responses/ProviderError"
        "500":
          $ref: "#/components/responses/ProviderError"
  /v1alpha1/policy-provider/validate-bundle:
    post:
      tags: [PolicyProvider]
      operationId: policyValidateBundle
      summary: Validate policy bundle metadata or content before activation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyBundleValidationRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyBundleValidationResponse"
        "400":
          $ref: "#/components/responses/ProviderError"
        "500":
          $ref: "#/components/responses/ProviderError"
...
components:
  responses:
    ProviderError:
      description: Structured provider error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProviderError"
  schemas:
    ObjectMeta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId:
          type: string
          description: Correlation ID generated by the control plane
        timestamp:
          type: string
          format: date-time
        tenantId:
          type: string
        projectId:
          type: string
        environment:
          type: string
        actor:
          type: object
          additionalProperties: true
    ResourceRef:
      type: object
      properties:
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
        uid:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
    SubjectRef:
      type: object
      properties:
        type:
          type: string
          description: user, serviceaccount, workload, system, tenant
        id:
          type: string
        attributes:
          type: object
          additionalProperties: true
    ActionRef:
      type: object
      required: [verb]
      properties:
        verb:
          type: string
        target:
          type: string
        attributes:
          type: object
          additionalProperties: true
    DecisionReason:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    Obligation:
      type: object
      required: [type]
      properties:
        type:
          type: string
        config:
          type: object
          additionalProperties: true
...
    ProviderError:
      type: object
      required: [errorCode, message]
      properties:
        errorCode:
          type: string
        message:
          type: string
        retryable:
          type: boolean
          default: false
        details:
          type: object
          additionalProperties: true
    PolicyEvaluateRequest:
      type: object
      required: [meta, subject, action]
      properties:
        meta:
          $ref: "#/components/schemas/ObjectMeta"
        profile:
          type: object
          properties:
            profileId:
              type: string
            profileVersion:
              type: string
            source:
              type: string
        subject:
          $ref: "#/components/schemas/SubjectRef"
        action:
          $ref: "#/components/schemas/ActionRef"
        resource:
          $ref: "#/components/schemas/ResourceRef"
        context:
          type: object
          additionalProperties: true
        mode:
          type: string
          enum: [enforce, audit]
          default: enforce
        dryRun:
          type: boolean
          default: false
    PolicyEvaluateResponse:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [ALLOW, DENY, CHALLENGE, DEFER]
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/DecisionReason"
        obligations:
          type: array
          items:
            $ref: "#/components/schemas/Obligation"
        policyBundle:
          type: object
          properties:
            policyId:
              type: string
            policyVersion:
              type: string
            checksum:
              type: string
        evidenceRefs:
          type: array
          items:
            type: string
        output:
          type: object
          additionalProperties: true
    PolicyBundleValidationRequest:
      type: object
      required: [meta, bundle]
      properties:
        meta:
          $ref: "#/components/schemas/ObjectMeta"
        bundle:
          type: object
          properties:
            policyId:
              type: string
            policyVersion:
              type: string
            contentType:
              type: string
            checksum:
              type: string
            uri:
              type: string
              format: uri
            inline:
              description: Optional inline bundle payload for small bundles
        expectedCapabilities:
          type: array
          items:
            type: string
    PolicyBundleValidationResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/DecisionReason"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/DecisionReason"
        discoveredCapabilities:
          type: array
          items:
            type: string
...